+++ src/test/java/com/project/savingbee/productAlert/ProductAlertServiceTest.java
@@ -10,6 +10,9 @@ import static org.mockito.Mockito.verify;
 import com.project.savingbee.common.entity.ProductAlertSetting;
 import com.project.savingbee.common.entity.ProductAlertSetting.AlertType;
 import com.project.savingbee.common.repository.ProductAlertSettingRepository;
+import com.project.savingbee.productAlert.dto.AlertSettingsRequestDto;
+import com.project.savingbee.productAlert.dto.AlertSettingsResponseDto;
+import com.project.savingbee.productAlert.service.ProductAlertService;
 import java.math.BigDecimal;
 import java.math.BigInteger;
 import java.time.LocalDateTime;
